ADDED ABOUT PAGE
Created About component folder for the page with index.jsx and .scss files

About page has a header, 3 paragraphs, and a custom animated spinning cube
where the cube faces have an image of 6 common web skills.

The page fades in with the header letters appearing sequentially, the
paragraphs are highlighted with a pulse animation, and the cube
immediately begins spinning.

The header uses the AnimatedLetters component for its effect and reuses
the useEffect and useState from the Home component header for the same
animation. Hovering changes the letters from yellow to white. header
tags encapsulate the text.

The paragraphs use a quick pulse effect and nth child selector to
quickly emphasize one after the other sequentially.

The cube loads and immediately begins spinning indefinitely using the
custom animation and with a glow thanks to box-shadow.
Each face contains the logo of a popular web technology.